chore(core): optimization
